refactor to encompass more than one route algo
